Add skeleton files
Added information on the README about what this
project is about and what it demonstrates.
Added a html file with skeleton code to have a
structure to put new code on. Added comments for
different sections of the desired outcome. Acts
as an outline for better workflow.
Added CSS to prepare for styling.