Add styling and spacing
Added flex on list of links, so I can have the
links line up in one row instead of one column.
Added padding to the sides of the header so it
looks closer to the desired outcome.
Added spacing between the links to separate them.
Overrided font weight of the links so they don't
get affected by the font weight of the logo.
Added weight to the logo as indicated by the
color and fonts style guideline.